Change NSIntegers to NSUIntegers

+++ CocoaMail/Views/UI/MailListViewController+UIViewControllerPreviewing.m
@@ -33,8 +33,8 @@
         UIStoryboard *storyboard = [UIStoryboard storyboardWithName:@"Main" bundle:nil];
         PreviewViewController *previewController = [storyboard instantiateViewControllerWithIdentifier:@"PreviewView"];
         
-        NSInteger dayIndex = path.section;
-        NSInteger conIndex = path.row;
+        NSUInteger dayIndex = (NSUInteger)path.section;
+        NSUInteger conIndex = (NSUInteger)path.row;
         ConversationIndex *ci = [self.convByDay conversation:conIndex onDay:dayIndex];
         Conversation* conv = [[Accounts sharedInstance] conversationForCI:ci];
